+++ diventi/landing/admin.py
@@ -1,6 +1,6 @@
 from django.contrib import admin
 
-from .models import Presentation, Feature, Event
+from .models import Presentation, Feature
 
 
 class FeatureInline(admin.TabularInline):
@@ -8,18 +8,10 @@ class FeatureInline(admin.TabularInline):
     fields = ('title', 'icon', 'color', 'description')
     extra = 0
 
-
-class EventInline(admin.TabularInline):
-    model = Event
-    fields = ('title', 'icon', 'color', 'description', 'event_date')
-    extra = 0
-
-
 class PresentationAdmin(admin.ModelAdmin):
     list_display = ['title', 'active']    
     inlines = [
         FeatureInline,
-        EventInline,
     ]
 
 admin.site.register(Presentation, PresentationAdmin)